File system-related system calls. (#13)
The goal of this PR is to implement file system-related system calls in `polkakernel`. The means of achieving this goal is to make CoreVM run most of the  `busybox`'s subcommands.

* Fix thread-local storage in `musl` to make it work with locales.
* Update `FileSystem` to include directory reading functions.
* Update `Kernel` to support FS-related system calls as well as common syscalls like `getcwd`, `uname` etc.
* Add a script to build `busybox`.

The following `busybox` commands are known to work: `ls`, `find`, `stat`, `du`, `pwd`, `cat`, `date`, `uname`, `id`, `hostname`, `arch`, plus almost all commands that read a file and do something with it (e.g. `sha256sum`, `wc`).

Notable commands that don't work:
- `stat -f` (requires either doing full FS traversal or implementing metadata storage in the root directory on the CoreVM side);
- `xxd` (requires stdin redirection which in turn needs sophisticated serialization of aliased FDs on the CoreVM side).